Split to separate job step
