+++ tests/actor_pool/test_actor_cuda_devices.py
@@ -20,35 +20,25 @@ pytestmark = pytest.mark.skipif(
 
 
 @contextmanager
-def reset_runner_with_gpus(num_gpus, monkeypatch):
+def reset_runner_with_gpus(num_gpus):
     """If current runner does not have enough GPUs, create a new runner with mocked GPU resources."""
     if len(cuda_visible_devices()) < num_gpus:
-        if get_tests_daft_runner_name() == "ray":
-            try:
-                ray.shutdown()
-                ray.init(num_gpus=num_gpus)
-                yield
-            finally:
-                ray.shutdown()
-                ray.init()
-        else:
-            try:
-                monkeypatch.setenv("CUDA_VISIBLE_DEVICES", ",".join(str(i) for i in range(num_gpus)))
-
-                # Need to reset runner to recompute resources
-                original_runner = daft.context.get_context()._runner
-                daft.context.get_context()._runner = None
-                yield
-            finally:
-                daft.context.get_context()._runner = original_runner
+        assert get_tests_daft_runner_name() == "ray"
+        try:
+            ray.shutdown()
+            ray.init(num_gpus=num_gpus)
+            yield
+        finally:
+            ray.shutdown()
+            ray.init()
     else:
         yield
 
 
 @pytest.mark.parametrize("concurrency", [1, 2])
 @pytest.mark.parametrize("num_gpus", [1, 2])
-def test_actor_pool_udf_cuda_env_var(monkeypatch, concurrency, num_gpus):
-    with reset_runner_with_gpus(concurrency * num_gpus, monkeypatch):
+def test_actor_pool_udf_cuda_env_var(concurrency, num_gpus):
+    with reset_runner_with_gpus(concurrency * num_gpus):
 
         @udf(return_dtype=DataType.string(), num_gpus=num_gpus)
         class GetCudaVisibleDevices:
@@ -79,8 +69,8 @@ def test_actor_pool_udf_cuda_env_var(monkeypatch, concurrency, num_gpus):
         assert len(all_devices) == concurrency * num_gpus
 
 
-def test_actor_pool_udf_fractional_gpu(monkeypatch):
-    with reset_runner_with_gpus(1, monkeypatch):
+def test_actor_pool_udf_fractional_gpu():
+    with reset_runner_with_gpus(1):
 
         @udf(return_dtype=DataType.string(), num_gpus=0.5)
         class FractionalGpuUdf:
